DJK-57: Admin-Accounts löschen, diverse Formulare, die nur 1x verwendet werden, wieder inline in der route. delete cascade für Spieler und Feedback ins Schema eingetragen.
