+++ app/routes/application/admin/users/$userId.tsx
@@ -1,0 +1,11 @@
+import {Outlet} from "@remix-run/react";
+
+const UserId = () => {
+    return (
+        <>
+            <Outlet/>
+        </>
+    )
+}
+
+export default UserId
