+++ app/routes/application/admin/users/invites/$inviteId.response.tsx
@@ -1,22 +1,22 @@
 import {ActionFunction, json, LoaderFunction, redirect} from "@remix-run/node";
 import {AdminInvitation} from "@prisma/client";
-import {useCatch, useLoaderData, useNavigate, useParams} from "@remix-run/react";
+import {Form, useCatch, useLoaderData, useNavigate, useParams} from "@remix-run/react";
 import {DateTime} from "luxon";
 import {getAdminInvitation, updateAdminInvitation} from "~/models/admin.user.invitation.server";
-import InviteChoiceForm from "~/components/users/admin/InviteChoiceForm";
 import routeLinks from "~/helpers/constants/routeLinks";
 import {Params} from "react-router";
 import DefaultButton from "~/components/common/buttons/DefaultButton";
 import messages from "~/components/i18n/messages";
 import React from "react";
+import PageHeader from "~/components/common/PageHeader";
+import InputWithLabel from "~/components/common/form/InputWithLabel";
+import dateUtils from "~/dateUtils";
+import ButtonContainer from "~/components/common/container/ButtonContainer";
+import RedButton from "~/components/common/buttons/RedButton";
+import SuccessButton from "~/components/common/buttons/SuccessButton";
 
 type LoaderData = {
     adminInvitation: AdminInvitation,
-}
-
-type AdminToken = {
-    email: string
-    expires_at: string
 }
 
 export const action: ActionFunction = async ({params, request}: { params: Params, request: Request }) => {
@@ -65,7 +65,45 @@ const InviteChoiceView = () => {
     const {adminInvitation} = useLoaderData() as LoaderData;
     const validUntil = DateTime.fromJSDate(new Date(adminInvitation.expires_at))
     return (
-        <InviteChoiceForm validUntil={validUntil} name={adminInvitation.name} email={adminInvitation.email}/>
+        <>
+            <header>
+                <PageHeader title={messages.adminInviteUserChoiceForm.title}/>
+            </header>
+            <Form method={"post"} className="md:w-1/2">
+                <main className={"flex flex-col gap-4"}>
+                    <InputWithLabel label={messages.adminInviteUserChoiceForm.email}
+                                    type='email'
+                                    name={'email'}
+                                    id={'email'}
+                                    disabled={true}
+                                    defaultValue={adminInvitation.email || ''}
+                    />
+                    <InputWithLabel label={messages.adminInviteUserChoiceForm.validUntil}
+                                    type='text'
+                                    name={'validUntil'}
+                                    id={'validUntil'}
+                                    disabled={true}
+                                    defaultValue={dateUtils.dateToFormat({value: validUntil})}
+                    />
+                    <InputWithLabel label={messages.adminInviteUserChoiceForm.name}
+                                    type='text'
+                                    name={'adminName'}
+                                    id={'adminName'}
+                                    defaultValue={adminInvitation.name}
+                    />
+                    <ButtonContainer>
+                        <RedButton>
+                            <button name='intent' value={'reject'}
+                                    type={'submit'}>{messages.adminInviteUserChoiceForm.rejectInvitation}</button>
+                        </RedButton>
+                        <SuccessButton className='ml-auto'>
+                            <button type={'submit'} name={'intent'}
+                                    value='accept'>{messages.adminInviteUserChoiceForm.acceptInvitation}</button>
+                        </SuccessButton>
+                    </ButtonContainer>
+                </main>
+            </Form>
+        </>
     )
 }
