+++ app/routes/application/admin/users/invites/new.tsx
@@ -2,7 +2,6 @@ import type {LoaderFunction} from "@remix-run/node";
 import {ActionFunction, json, redirect} from "@remix-run/node";
 import {requireUserId} from "~/session.server";
 import invariant from "tiny-invariant";
-import InviteUserForm from "~/components/users/admin/InviteUserForm";
 import dateUtils from "~/dateUtils";
 import {DateTime} from "luxon";
 import {createEncryptedAdminToken} from "~/utils/token.server";
@@ -11,7 +10,16 @@ import {adminInvationMail} from "~/components/i18n/adminInvationMail";
 import mailLinkBuilder from "~/helpers/mail/mailLinkBuilder";
 import mailSender from "~/helpers/mail/mailsender";
 import ErrorView from "~/components/errorhandling/ErrorView";
-import {Outlet} from "@remix-run/react";
+import {Form, Outlet} from "@remix-run/react";
+import PageHeader from "~/components/common/PageHeader";
+import messages from "~/components/i18n/messages";
+import InputWithLabel from "~/components/common/form/InputWithLabel";
+import {DateInput, formatForDatePicker} from "~/components/common/datetime/datetime";
+import ButtonContainer from "~/components/common/container/ButtonContainer";
+import RedButton from "~/components/common/buttons/RedButton";
+import DefaultButton from "~/components/common/buttons/DefaultButton";
+import {useState} from "react";
+import routeLinks from "~/helpers/constants/routeLinks";
 
 
 export const loader: LoaderFunction = async ({request}: { request: Request }) => {
@@ -63,7 +71,7 @@ export const action: ActionFunction = async ({request}: { request: Request }) =>
         const mailTo = email
         const body = adminInvationMail.mailBody({expiresAt: validUntilTxt, adminName, einladungsLink})
         await mailSender.sendMail({mailTo, subject, body})
-        return json({data: {invitation}})
+        return redirect(routeLinks.admin.users.home)
     } catch (error) {
         console.log("ERROR SENDING MAIL", error)
         return json({data: error}, 500)
@@ -71,9 +79,58 @@ export const action: ActionFunction = async ({request}: { request: Request }) =>
 }
 
 export const NewAdmin = () => {
+    const [dateValue, setDateValue] = useState<DateTime>(DateTime.now().endOf('day').plus({day: 7}))
+
+    const onDateValueSelect = (value: string) => {
+        const datePickerValue = dateUtils.dateFromFormat({text: value, options: {format: dateUtils.datePickerFormat}})
+        setDateValue(dateValue.set({
+            year: datePickerValue.year,
+            month: datePickerValue.month,
+            day: datePickerValue.day
+        }))
+    }
+
     return (
         <>
-            <InviteUserForm/>
+            <header>
+                <PageHeader title={messages.adminInviteUserForm.title}/>
+            </header>
+            <Form method={"post"} className="w-1/2">
+                <main className={"flex flex-col gap-4"}>
+                    <InputWithLabel label={messages.adminInviteUserForm.email}
+                                    type='email'
+                                    name={'email'}
+                                    id={'email'}
+                                    defaultValue={''}
+                    />
+                    <InputWithLabel label={messages.adminInviteUserForm.name}
+                                    type='text'
+                                    name={'adminName'}
+                                    id={'adminName'}
+                    />
+                    <DateInput onChange={onDateValueSelect}
+                               value={formatForDatePicker(dateValue)}
+                               name='validUntil'
+                               label={messages.adminInviteUserForm.validUntil}
+                    />
+                    <InputWithLabel label={messages.adminInviteUserForm.scope}
+                                    type='text'
+                                    name={'scope'}
+                                    id={'scope'}
+                                    defaultValue='admin'
+                    />
+
+                    <ButtonContainer>
+                        <RedButton>
+                            <button name='intent' value={'cancel'} type={'submit'}>{messages.buttons.cancel}</button>
+                        </RedButton>
+                        <DefaultButton className='ml-auto'>
+                            <button type={'submit'} name={'intent'} value='save'>{messages.buttons.save}</button>
+                        </DefaultButton>
+                    </ButtonContainer>
+
+                </main>
+            </Form>
             <Outlet/>
         </>
     )
